Cut any food in array & fixed error non cuttable
Cut any food that is in the array  (Tomato, Cabbage, Cheese) and fixed error when trying to cut a non 'cuttable' food. Error occurred when trying to cut something that can't be cut e.g. bread, sliced food that has already been cut. Trying to cut resulted in a game break